Remove duplicated BLAS rewriting code
Accidentally introduced in c655b0285ddd2c100bd260fd1e581e9376be2553

Also move tests to the rewriting test file
